up k

+++ CoreBase/Test/Dictionary/frmBranch.cs
@@ -236,8 +236,22 @@ namespace AusNail.Dictionary
             if (result == DialogResult.Yes)
             {
                 this.zDeleteRow = ((DataTable)Bds.DataSource).Rows[_postion];
-                bool flag = base.DeleteData();
-                LoadData();
+                // Check use for bill or booking.
+                DataTable dataTable = MsSqlHelper.ExecuteDataTable(ZenDatabase.ConnectionString, "zBranchCheckUse", int.Parse(zDeleteRow["branchId"].ToString()));
+                if (dataTable != null && dataTable.Rows.Count > 0)
+                {
+                    MessageBox.Show("Branch is already in use, cannot be deleted.!", "Error", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                    return;
+                }
+                else
+                {
+                    bool flag = base.DeleteData();
+                    LoadData();
+                }
+                
+
+                //bool flag = base.DeleteData();
+                //LoadData();
             }
             else
             {
